crypto/little_e/requirements.txt
gmpy2==2.1.5
pwntools==4.10.0
pycryptodome==3.18.0
